requirements.txt
discord
python-dotenv
requests
pytz